# Conflicts:
#	CHANGELOG.md

+++ src/Console/KeysCommand.php
@@ -3,7 +3,9 @@
 namespace Laravel\Passport\Console;
 
 use Illuminate\Console\Command;
+use Illuminate\Support\Arr;
 use Laravel\Passport\Passport;
+use phpseclib\Crypt\RSA as LegacyRSA;
 use phpseclib3\Crypt\RSA;
 
 class KeysCommand extends Command
@@ -39,10 +41,17 @@ class KeysCommand extends Command
         if ((file_exists($publicKey) || file_exists($privateKey)) && ! $this->option('force')) {
             $this->error('Encryption keys already exist. Use the --force option to overwrite them.');
         } else {
-            $key = RSA::createKey($this->input ? (int) $this->option('length') : 4096);
+            if (class_exists(LegacyRSA::class)) {
+                $keys = (new LegacyRSA)->createKey($this->input ? (int) $this->option('length') : 4096);
 
-            file_put_contents($publicKey, (string) $key->getPublicKey());
-            file_put_contents($privateKey, (string) $key);
+                file_put_contents($publicKey, Arr::get($keys, 'publickey'));
+                file_put_contents($privateKey, Arr::get($keys, 'privatekey'));
+            } else {
+                $key = RSA::createKey($this->input ? (int) $this->option('length') : 4096);
+
+                file_put_contents($publicKey, (string) $key->getPublicKey());
+                file_put_contents($privateKey, (string) $key);
+            }
 
             $this->info('Encryption keys generated successfully.');
         }
